Check whether you can pray before recording that you do.
When the #pray command is issued, un the can_pray check before
recording the prayer in the history and breaking atheist conduct.

This is on the whole a quite minor bugfix.  The most notable known
case is when you are polyselfed into a demon and the idea of praying
to a non-chaotic god is so repugnant that your character outright
refuses to do so.  Previously, the history would nonetheless report
that you did, and conduct would be violated.  Now it won't.

Thanks to FIQ for pointing out this deficiency.