Polearms now get their own status hilight, similar to ranged weapons.
It's not really fair to call them non-weapons, or treat wielding one
as being unarmed; but they're not melee weapons either.  This
introduced the "Pole" status hilight.

Also, I renamed "cWielded", which I never liked because A) it's not a
word, and B) it looks like a variable name, not a human-readable word.
The reason it existed was because someone objected that "Stone" or
"Petrify" might make the user think they were turning to stone.  So I
went with "Trice" this time, because it's shorter than "RubberChicken".