Make the Sceptre grant drain res, and fix some quest text bugs.
The Priest quest artifact, the Sceptre of Might, now grants drain
resistance when wielded.  Admittedly, this will be anticlimactic,
since priests must fight through hordes of undead (including some with
draining attacks) in order to get the artifact in the first place.
But it's thematically appropriate, and it makes a quest text message
make sense that otherwise wouldn't.

Speaking of which, fix up some priest quest text messages that
referred to wearing the quest artifact, left over from when it was a
helm.  Thanks to K2 for noticing these.