When you see a non-blank 60zm scroll in a shop, type-identify it.
The general-purpose code for type-identifying unambiguous price/weight
item combinations in shops wasn't picking up the scroll of enchant
weapon, because the blank scroll is the same price and weight.  But
blank scrolls aren't random-appearance, so there's no actual ambiguity.

However, in looking at this, I noticed that the general-purpose code
ignores the possibility of surcharges.  I put a note in TODO about
that but haven't fixed it yet.