Split off Unicode tileset, with dragon diacritics, from legacy tileset, without.
It has been brought to my attention that there are people still using
terminals that apparently haven't been updated since before the Gnome
team stopped adding features and started removing them, way back when
Unicode support was considered an optional extra; but they aren't
using the ASCII tileset; libuncursed is doing some magic to detect the
situation and support their terminals, as long as all of the
characters in the tileset are representable.  (If not, they get
something else, upside-down question marks in my testing, instead of
the intended character.)  This is apparently also relevant when using
a text-based tileset in SDL mode (not that there's any real reason to
do that, other than testing).

In order to accommodate the situation, there are now three text-based
tilesets, rather than just two.  They are as follows:

 1. ASCII - state of the art 1980.  (I can get away with saying this,
    because personally this is the tileset that I usually use.)

 2. Legacy/CP437 - state of the art 1995.  Replaces both IBMGRAPHICS
    and DECGRAPHICS, but does not use e.g. consonants with diacritics.
    This is the tileset formerly called "Unicode", but it is in fact
    much more limited than that.

 3. Unicode - requires your terminal to actually speak UTF-8.  This
    allows e.g. all five different levels of dragons to be visually
    distinguishable without farlook, using diacritcs.

If you see upside-down question marks in dragon halls, or other
evidence of unsupported characters, consider switching your tileset
from Unicode to Legacy/CP437.