Make shop unique-price-weight combo auto-ID take surcharges into account.
When you see an item on a shop floor, with its price, the game
calculates whether its price is unique (given its item class and
weight, which are known) in order to decide whether to automatically
mark the item type as known (because it can be inferred from price).

Previously, this didn't take the possibility of surcharges into
account in any way.

Now, if the seen item is surcharged, it's checked against items whose
base price matches its surcharged price; whereas, if the seen item is
not surcharged, it's checked against items whose surcharged price
matches its base price.  (In both cases, it's still also checked
against items of its own price.)

I didn't go through the item list to see whether this is currently
relevant for any specific item.  The list of objects and their prices
are subject to change, and the code should handle that, barring
special/unusual cases (such as the inclusion of a fixed-appearance
item in a random-appearance price bracket, as with the 60zm scroll,
which is special-cased).