Fix the incorrect plural of poisoned ya.
Apparently, marking a plurals_dictionary entry with ^ means it has to
match the entire thing being pluralized.  The comment says it only has
to match the whole word, but maybe it assumes that the function is
being handed a word, and "poisoned ya" was getting passed to it as a
single entity.