Creektown Overhaul: use a single MAP section.
I understand why the original author of Creektown built the map out of
multiple sections.  It allows for a non-rectangular shape for the
designed portion of the level, allowing the portions around the
corners to be randomly generated.

However, it has been leading to (or perhaps exposing) bugs.  The level
doesn't always get pieced together exactly correctly.

What I _want_, ultimately, is to be able to specify portions of the
MAP as a flexible to-be-decided terrain (perhaps represented by ? in
the MAP specification), so that the level can be specified as a single
map and yet allow for non-rectangular designs, with the remainder of
the level generated procedurally.  I believe that this will be MORE
flexible in the long run than the multi-MAP functionality, and I also
believe it will be easier for level designers and maintainers to keep
track of what's going on with the design of the level.

Ad interim, I just went with a fixed design for this level for now;
there are two fixed-map Minetown variants already, so a third is not a
huge disaster in the short term.  Better than having abandoned shops
because the keeper doesn't recognize one of the walls, and other such
bugs that have been occurring.