Close weird loophole in the alignment record conduct.
In certain cases (which are fairly uncommon in practice) it's possible
to get your alignment record directly set to a particular value,
rather than having points added to or subtracted from it.  Previously,
the lostalign conduct was not updated if the net effect of this was a
decrease.  Now it is.