tiletool: add support for large-palette tilesets (e.g., slashem-32).
The following features of the slashem-32 tileset were previously
unsupported by the tiletool but are supported now:
 * two-character palette keys
 * distinct palettes in each source file

For the time being, these features are bundled together, i.e., it is
assumed that a tileset that uses one of them uses the other also.  It
would be possible to decouple them, and possibly even to allow some
source files to use single-character palette keys while others use
double-character palette keys; but those enhancements aren't needed
for the slashem-32 tileset, so they haven't been implemented yet.