Initial support for building (incomplete) slashem-32 tileset.
The 32-pixel-square 32-bit-color tileset from Slash'EM currently isn't
usable, because it only has the tiles that Slash'EM uses and thus is
missing a significant number that Fourk would need.

However, the next step to fixing that (now that we have tiletool
suppport for it) is to add the ability to build it.  The tileset
compiler will now tell us, at build time, a list of tiles still
needed.