Redo (by hand) broken merge of function fill_dragonhall()
This function had significant changes in both the master and
bleeding-edge branches, for different reasons.  The first attempt at
the merge, I let git merge do what it could and then just looked at
the places it thought it couldn't and attempted to merge those parts
manually.  This resulted in an incorrect merge that wouldn't even
compile (there were duplicate case numbers in a switch statement).
Rather than attempt to figure out how the first merge went wrong, I
just grabbed the master version of the function and then manually
re-applied the changes to that function from commit f7bf73112c, which
added yellow and black dragon halls, dragon eggs in dragon halls, and
adjusted the appearance themed loot to account for the fact that there
are now spare appearances for rings, books, and potions, so any given
appearance may not be used in any given game.