Experimental bugfix: selling a non-empty container for hardware store credit.
Tangled noticed this one a while ago:  when you sell a non-empty
container in a hardware store and are paid in gold, you retain
ownership of its non-hardware-relevant contents.  However, if you
sold it for shop credit, then you relinquished ownership of all the
contents, even though you only received credit for hardware items.

Perhaps the correct solution is to declare that "shopkeepers are
greedy" and have them always pay for containers as if empty, and
assume ownership of all the contents.  That would be consistent.
But I haven't implemented that.

Instead, I made the shop-credit case call dropped_container, like
the sell-for-gold case does.  So now if you sell a sack of potions
for shop credit, you get 1zm of credit, but the shopkeeper doesn't
take ownership of the potions, the same as if you sold for gold.