Don't give scurrier rangers a single +1 dart in starting inventory.
This happened because the starting +1 dagger was converted to a dart.
Adjust its enchantment to +2 in that case so that it can stack with
the others.  Thanks to rumflump for pointing this out.

There's still a corner case wherein a scurrier ranger can end up with
a single blessed +2 dart, and the rest uncursed (or possibly vice
versa).  Fixing that would be more awkward unless we want to guarantee
the beatitude of all the +2 darts, so leave that for now.