For sacrifice gifts, try not to give size-inappropriate armor.
Currently, this is going to be imperfect, because some playable races
have so few armor options, especially in certain slots, that the gift
code will not be able to choose anything that fits.  The proper fix to
this is to add fallback options that _will_ fit, but that will have to
wait for 4.3.0.5, when I plan to add things like Gnomish body armor.