Don't give inappropriately-racial armor as a sacrifice gift.
When giving armor as a sacrifice gift (which is a bit of a corner
case, since it only happens in case of conduct play), don't give elven
armor to an orcish player, orcish armor to a dwarven player, etc.

Also, try harder not to give something that doesn't fit, by setting
the default to something that does.  (The default default is a
t-shirt, but if that is too large or small, default to a small or
large shield, instead.)