Refactor player movement rate calculation, and show current speed in status area.
The player's current speed is now calculated by a function,
you_speed(), so that it can be called in more than one place; the botl
code now uses this to show the player's current movement rate (M:
because S: is score) in the status area, and it is also used by the
(also new) function you_move_amount(), which calculates how many
movement points the player is getting on this specific turn, which is
where the random component of player movement rate is calculated.  The
place in the you_moved() function that used to do all the player speed
calculations now just calls this function you_move_amount() to find
out how many movement points to give the player.

I kind of want to make you_speed() just call the monster speed
function, mcalcmove(), with &youmonst, and put the speed calculations
there, to ensure monster/player symetry; but that would require some
additional refactoring, for a couple of reasons (e.g., currently
mcalcmove returns u.moveamt if called with &youmonst, and we'd have to
change that, and I would want to track down all the places where
mcalcmove is called and make sure this wouldn't be a problem).  So
that aspect of the refactor is postponed for now.