Bugfix: set generating_dump when generating a dumplog.
Thanks to Mandevil for noticing that this was not happening.  (The
oversight was typically masked by the fact that answering yes to the
DYWYPI prompt set it in order to show the identities of items, and
this happened prior to generating the dump.  So the problem only
showed up when someone answered no to DYWYPI and then looked at the
dumplog.)