Revert "Refactor player movement rate calculation, and show current speed in status area."
This reverts commit 8a6580c069cad23696a5e466665e0b3dcf9b63b7.

The status area part was fine, but when I refactored the speed
calucation out to a function, I did something wrong.  Very wrong.
Thanks to FIQ for noticing this before release, because this would've
been a REALLY embarrassing bug to have in an actual release, as it
would likely make the game effectively unplayable.  The rate at which
movement was doled out was very very wrong, giving you increasingly
many actions per turn until suddenly, boom, 12 turns of none.

Conflicts:
	doc/changelog-fourk.txt
	doc/todo-fourk.txt