Further player speed fixups: use an increment smaller than NORMAL_SPEED.
Adjust the player movement point calculation so that it is slightly
less irregular, particularly when the player is a little slower than
NORMAL_SPEED.  Instead of doling out movement points only in
increments of NORMAL_SPEED, we now dole them out in increments that
are a fraction of that.  This means that for example if you are speed
nine, it's still not predictable which actions will take two turns and
which will take only one turn, but no single-turn action will take
three turns.  As before, the long term average should converge toward
the expected value.

Into the bargain, even NORMAL_SPEED player characters do get a little
bit of unpredictability: they will occasionally miss a turn or,
conversely, get a second action on the same turn.