Temporarily disable the alternate castles with problematic mazewalks.
When I first imported these two levels on bleeding-edge, they
exhibited problems with mazewalk, which turned out to be due to
even/odd coordinate parity.  I adjusted them to make them work
correctly, but I did so on the bleeding-edge branch, which I normally
build with larger-than-default ROWNO/COLNO values (so that I can see
the bugs that happen as a result of this, such as the vanilla Castle
being trivially bypassed around the north or south side of the maze,
Creektown in its original multi-part incarnation not being pieced
together correctly, etc.).  Later, when the code became save stable, I
merged it to the master branch, which is currently still compiled with
traditional default values for ROWNO/COLNO (precisely so that users on
public servers do NOT encounter the aforementioned class of bugs,
until I am ready to actually fix them).  Unfortunately, I didn't
realize, this shifted the parity again, so that now the levels I had
fixed were once more incorrect.  (The vanilla Castle and the Catacombs
version of Mine's End do not have this problem, perhaps because of a
difference in even/odd parity in the total number of rows and columns
in each of the maps.  It's complicated.)

The correct solution is to fix mazewalk so it doesn't care about even
versus odd coordinates relative to the whole level (ROWNO/COLNO), only
relative to the starting position.

Until then, don't use the levels that exhibit the problem for now.