Fix a "may be used uninitialized" warning by pre-initializing.
This *should* never be relevant, as the loop that conditionally does
the initialization should always have at least one valid case it can
pick.  (I suppose somebody could find a way to unlock a race that
can't be any of the currently-unlocked roles, though I don't know of a
specific way to do that, and set that race a preference...)  In any
case, we pre-initialize to 1, which I believe should be Barbarian
based on the order in role.c.  If the situation does come up, this
will likely still leave a bug, but at least it won't be using
uninitialized memory.