Prepare infrastructure for release preparation (last-updated comments).
During development, I distribute the code only via git repository,
which, as a core part of its functionality, provides information such
as who last edited each file (or each line of each file, for that
matter) and when, and what they said about their edit at the time.  I
consider this to substantially exceed, at least in spirit, the NGPL
license requirement to provide that information.

However, I am considering the possibility, with the 4.3.0.4 release,
of possibly distributing via other mechanisms, and if so it would be
necessary to provide that information in some other way; the most
obvious way to do this is with comments at the top of each file, as
was historically done.

This commit, therefore, introduces a script (update-last-updated.pl)
that iterates over the codebase and asks git for that information, in
order to modify the comments at the top.

While testing this script, I discovered that some of the files do not
have such a comment or fail to explicitly mention their license
status.  So in addition to adding the script, this commit corrects
those omissions, so that the script will be able to update everything
it needs to update.

Note that currently the script does a dry run:  it doesn't actually
make the changes.  This will be trivial to change when I am ready to
actually use the script, however.

Also note that actually using this script should not be done in the
repository itself, on a branch where further development will take
place, or which will ever be merged.  (Doing so would cause every
future merge or cherry-pick to have merge conflicts on every single
file that was touched.)