Minor buffs for Samurai role and Giant race.
Samurai can now get expert in polearms, and receive a halberd in the
magic chest and a glaive on the quest.

Giants now start with ten extra points of Defense, compared to other
races.  This makes very good sense if you think in terms of damage
reduction, and the opposite of sense if you think in terms of
evasiveness and the difficulty of hitting them.  I want to make Arm
and Ev distinct properties, at which point this issue will be
revisited and the various races each will get different numbers.

But ad interim, giants really needed a buff, to offset their terrible
speed and inability to wear some important armors.