Humans sometimes get a potion of oil in monster starting inventory.
From a flavor perspective, most of the human monsters have equipment
of one kind or another (weapons, armor, whatever) that they could
plausibly need to maintain, so carrying oil makes sense for them.

The real reason, of course, is because I want potions of oil to be a
bit more available to the player.