Add icon for DLB, the data librarian.
Based on public domain clipart from the Open Clip Art Library, this is
not the most wonderfully artistic icon ever; but it should be
distinguishable from the other programs' icons and, at reasonable
sizes, should be mnemonic for the basic idea of what the program does
(not sure about 16x16 in that regard).  So an improvement over not
having an icon at all.

(cherry picked from commit 23c0e26ab6cc91a4ff986feda766708d7afe7379)