Let giants wield two-handed weapons in one hand.
Rather than hardcoding it for giants specifically, I check whether you
(in your current form) are (at least) huge.  Giants are currently the
only playable race that's huge in natural form, but a player who
starts as a dwarf or whatever can also do it if polyselfed into
something huge (or larger) that can wield weapons, e.g., a giant.

You can combine this with a shield, but not with two-weapon combat.