Remove meaningless "score" number from status area.
This isn't useful information.  It wouldn't be useful even if it were
somehow accurate, which it isn't.  I was going to make showing it an
option, off by default, but apparently the options are not in scope in
the function that draws the status area.  That could probably be
worked around, but it isn't worth it unless someone can make a
coherent argument that some players might have a better experience
with it turned on.  (It would have to be an argument that doesn't boil
down to "vanilla shows a score", because frankly, the vanilla score
is, despite being more accurate, nonetheless even more meaningless
than the completely inaccurate predictor of score that we had in
Fourk, because vanilla score doesn't reflect anything at all.)