Don't assume thrown potions target the player
It's possible for a monster to select another monster to aim the
potion at. Previously, it was using the range to the player to
determine this, or crashing if the monster couldn't determine the
range. Given that it's entirely valid for a monster that isn't
aware that the player's around to throw a potion at another monster
out-of-sight, this code needed to be revised.

Ironically, the revised version is probably simpler than the
original.

Fixes #959.

(cherry picked from commit 1f7412a1ba31043dfa7cb6674307d054e544a393)

Conflicts:
	libnethack/src/muse.c