Creektown:  add additional ways to get from side A to side B.
Some players may kill the quantum mechanics.  Which is fair.  They
should be allowed to do that: they're hostile, and this game
encourages you to kill hostile monsters.  And then, if they didn't
have a way to dig, the player would potentially be stuck on the west
side of the map for possibly an annoyingly long period of time.  So we
introduce an alternative way to get back, in the form of traps.  This
level didn't have very many traps anyway:  just one, in fact.  Now it
has several, and in keeping with its theme, most of them are
teleportation traps.