Shapechangers hold level-appropriate form longer, out-of-depth only briefly.
The random 1/6 chance per turn for a shapechanger to switch to a new
form is now 1/n, where in represents how depth-appropriate the form is
to the current dungeon level.  (The player's experience is NOT taken
into account, only the dungeon depth.)

Werefoo however still use the same 1/6 chance as before, for now.