Make lycanthropy trigger much less often, especially in daytime.
The full moon is also now relevant for this.  Worst case scenario, at
night during a full moon, the chance of lycanthropy transforming the
player into an animal is 1/100 each turn, which is what it was in the
daytime in vanilla.  Best case scenario, in the daytime when the moon
is not full, the odds are now 1/1000 each turn.

Lycanthropy is still a hazard, because it is unpredictable and can
trigger at quite inconvenient times.  But now, if you survive it the
transformation back to human, you may potentially be able to do things
like pick up your equipment, get it all put back on, and _walk to a
more tactically defensible position, possibly even on a different
dungeon level_, before it triggers again.  After several times, it may
even be possible to make it to a place where you have e.g. a stash of
holy water, or where there is an altar you can use to make it safe to
pray, or other ways you can solve your little problem.

Also, I decreased the difficulty numbers for a couple of wimpy
monsters (quasits, manes) to allow them to generate a little bit
earlier.

(cherry picked from commit 9550f7b4677a27128a1eee35c3065a660b8a1c3d)

Conflicts:
	doc/changelog-fourk.txt