Track how much gold is generated in various ways.
The game now tracks how much gold is generated in monster inventory,
on the floor (excluding vaults), in vaults (including the big one at
Fort Ludios), buried in stone, in containers, and by dungeon features
(fountains, thrones).  These data are recorded in the xlogfile.

My intention is to eventually use these data in-game as a way of
limiting total gold production, by having target figures for each
category and tapering each category off toward zero as it gets near to
its target value.  But that's for the future.  All I've implemented at
this point is the tracking.