Bugfix: segfault pertaining to recent spellbook generation changes.
Apparently in some situations, it is safe to call ledger_no(&lev->z)
but unsafe to call depth(&lev->z).  I don't understand why, but this
change allows the save to be loaded without a segfault, so this is
what I'm going with for the moment.  I may ask someone else later, to
see if I can get an explanation of what is actually going on here.  I
suspect it has to do with levels not being properly initialized at
certain points in the code flow, but I don't understand the details.