Thrown weapon appraisal:  appraise the entire stack.
When a stackable thrown weapon's enchantment is discovered due to
player character weapon skill when the weapon successfully hits a
monster, if the weapon was thrown from a stack, appraise the stack.

In case of multi-fire, when the appraisal doesn't happen on the first
throw, this still potentially leaves a small number of items
unappraised that should be appraised, i.e., the ones that were
previously thrown from the same stack.  This would be much harder to
fix, so I have left it for now.

Thanks to LarienTelrunya for discovering and reporting this case.