Spoiler:  special handling for M1_NOLIMBS
It turns out that it's not enough to check whether (mon->mflags1 &
M1_NOLIMBS) is true.  Because it'll be true if the monster has
M1_NOLIMBS, but it'll also be true if the monster merely has
M1_NOHANDS.  These two flags are stored in two bits, but in a
non-independent fashion.  This weirdness is inherited from vanilla,
and I have not changed it (though I was tempted).  I just fixed the
spoiler to test more precisely.