Don't try to free autopickup rules memory when there isn't any.
If the user has gone into the options and deleted all of their
autopickup rules, nethack4 would segfault any time the user's options
file was read, because it was calling free on a null pointer.

We now check that the pointer is non-null before calling free on it.

Thanks to AmyBSOD for reporting this bug.