If a moving monster is at position (COLNO,ROWNO), impossible() then relocate it.
If the player chooses to try to continue the game in this situation,
previously the game couldn't really be continued in practice, because
you would keep getting at least one impossible() per turn, maybe more.
Work around this by moving the monster to a valid location.

Ideally, we would want to track down how a monster got to such a
location in the first place, but that is harder.  For now, this will
let the player continue the game.