When deleting an engraving, don't impossible and fail to deallocate it.
Due to (what I take for) a typo, whenever del_engr() was called on an
engraving that wasn't the first one on the level's linked list of
engravings, the game first looped until a particular variable was
null, and then checked that variable to see whether to impossible
rather than actually deallocate the engraving.  It should have been
checking the other, very-similarly-spelled variable, that held the
engraving to be deallocated and would normally NOT be null.

Thanks to Luxcalibur on RLDiscord for reporting this bug.