Revert "When deleting an engraving, don't impossible and fail to deallocate it."
This reverts commit 89a70460787802bf7f042d575a5f7e0bdb9fc215.
I completely missed the fact that there's a break statement in
there, so that if the engraving being looked for is found, the
loop exits.  So the error condition only arises if the engraving
is not found on the level.

This means Fourk still has a bug, which occasionally causes del_engr()
to be called on an engraving that is not on the current level.