Attempt to be a little more robust when a monster carries a timed object onto a level.
Previously, this was just always a panic.  Now, if we can find a timer
on the current level's timers list that appears to relate to the
object in question, we set the object's olev to the current level and
pray that it works as intended.

This seems like it shouldn't happen (because how can the level have a
timer for an object that isn't already on the level, that seems bad
and wrong), and so this fix should do nothing, or at least nothing
good.  However, at least in the case of the save game that was sent to
me, this appears to work, and the game appears to then be continuable,
knock on wood.

The correct fix is probably to go back and prevent the situation from
happening in the first place.  But I don't know how to do that.