Adjust commit dab4b6f (carried object timer mitigation): cast other pointer
Per suggestion from ais523, cast the other pointer when doing this comparison,
since we don't know that the argument is a struct obj, and assuming that it is
"may cause some compilers to apply incorrect optimisations".

As before, this fix is still not the actual correct fix, which would be to
track dwon why the situation arises in the first place and prevent it.  Also
it may not cover all cases, particularly in the case of an object with more
than one timer.

But it still works with the test case we had, so I'm going with it for now.