Yet Another Attempt to fix pudding splitting and blood pudding generation.
As before, the intention is that blood puddings should only occur when
you've been splitting a pudding _repeatedly_, never on the first couple
of hits on a fresh brown or black pudding.

This time we're using an actual split counter, rather than relying on
the monster's other data to infer whether it's been split before.