Add struct_type field to various structs.
This feature is intended to make debugging certain types of
errors easier, by disambiguating between a situation wherein
the data fields of a struct have very incorrect data, versus
a situation wherein the pointer to the struct is stale or
points to the wrong kind of thing altogether.  I've wished
for the ability to distinguish these cases numerous times in
the past; doing it now because it breaks save compatibility,
and this is the correct time in my dev cycle to do that.