TODO list triage:  push some stuff back out of the .5 list.
If we want to keep alive the dream of ever getting 4.3.0.5 onto
the master branch, we need to compromise on some things.  So I
have now pushed a bunch of stuff out to a 4.3.0.6 list.  In
general, the thinking now is that unless it's some simple easy
thing, it's not needed for 4.3.0.5 unless it's critical for save
stability or likely to be needed to fix important bugs.