Bugfix:  frequent impossible() in the random number generator.
This was happening because of a change that generates random numbers
pre-emptively at the beginnings of certain functions (uhitm, mhitu,
etc.); in some cases, the arguments to these functions can be such
that there are no possible outcomes for the random number.  Don't
call the random number function in those cases.