Consolidate snakes, and add M3_CLIMBER flag.
This sets up the M3_CLIMBER flag, and adds it to the relevant monsters.
It currently doesn't *do* anything yet, but it's defined.

And we pare down the redundancy in mlet S.  This paves the way for
merging mlet N into S in some future release.

Commit also includes some further TODO list triage and a couple more
hallucinatory monsters, including the poster child for 2020.