Track and limit how many divine gifts you can get from sacrifice.
The game now tracks how many artifiact gifts and how many total
gifts you have received from sacrifice.  The number of artifact
gifts is limited to 3 in most modes, 1 in challenge mode.  For
now, the total number of gifts is not limited, so you can
continue to receive other objects even if you have reached the
limit on artifacts.

If you would have received an artifact, but you're maxed out on
artifact gifts, you may receive food, holy water, a random
amulet, or one of several potentially useful tools.  Note that
if you receive an amulet, it is not guaranteed to be one of the
positive/useful/safe ones.