Simplify the collection of player role names for consistency.
Previously, player characters (including the player, but also
player-character monsters) were referred to by a baroque collection
of names.  These were deterministic based on race, role, and gender;
but they were highly arbitrary and indicative of mostly nothing.
Most of the names in the list could have been switched to a
different role and/or a completely different experience level,
without seeming any stranger than they already were.

This commit simplifies the system:  we now use only the role and
gender to determine what to call the character, so e.g. instead of
"Canoness" or "Matriarch" or whatever, we now just say "Priestess"
and have done.

Some of the quest text has been updated to accommodate this change.