Level sound tracking and sound-inferred items on overview.
The game now tracks which "level sounds" (and smells, etc.) the player
has encountered on any given level.  In some cases, the level overview
can infer the existence of a level feature from this, even if the
player has not seen it.

There is likely more that can be done with this (e.g., the overview
currently does not mention the Oracle), but this will do for a start.
Having the tracking in place is the part that is needed before the
save compatibility freeze can happen.