Bugfix: incompatible pointer type when handling timers in mon_arrive
The code appeared to work before, and the warning was getting lost in
a long pile of warnings.  But now that I've cleaned up most of the
missing-tileset warnings and several other warnings that were due
to compiler updates and merges and various other issues, this one
emerged as something that probably needed to be fixed.  Actually,
I'm not entirely sure how the code could have been working before.