Overview: show vaults inferred from level sounds.
Now that we track which level sounds have been heard, this
is a relatively easy add.  Which is why I implemented tracking
level sounds in the first place.

Also, while we're at it, handle inferred altars slightly better:
use generic wording about the diety.  Rather than just saying
"an altar to " and then nothing, it now says an unknown god.