Price change: scroll of fire
The scroll of fire is now worth 200zm, the same base price as the
scrolls of water and earth.

The 100zm category still contains six scrolls, including both
useful and dangerous ones.  With proper preparation it is possible
to safely read-identify known-uncursed scrolls in this category,
but that was already true even with scroll of fire in the category,
and it's wasteful, so I don't think this is a big balance change.