Add pegasus tiles to Slash'EM, Dawnlike, and FMOB tilesets.
I don't know if I will get around to adding the pegasus to
this version of Fourk, but might as well be prepared.

Also, correct a couple of minor issues with the FMOB set
(e.g., one of the eyes of the owlbear had a transparent pixel
where it should have been opaque).

Also, add several bear (black, brown, grizzly, dire) tiles
to the Slash'EM 32 tileset.  I've wanted to add bears to the
game for a while; the lack of tiles, and the difficulty of
drawing them, has been the main thing holding me back.  This
is just about the easiest tileset to do them for, but at
least it's a start.