Don't make statues for non-birthcount cockatrices.
For example, if generating a temporary cockatrice monster for a
statue trap, don't create a bunch of additional statues around it.

Doing so could potentially lead to deeper recursion than is wanted,
at level-generation time.  In one instance, I experienced a delay
so long, I thought the game had gone into an infinite loop.  (The
recursion in that case was exacerbated by a larger than usual
number of iterations when choosing a nearby location for each
monster or statue being generated, since the room contained water
already, and with the addition of a couple of statue traps, the
area became quite crowded.)

In any case, if the monster's birth isn't being counted, it's
generally because the monster is temporary or otherwise quite
special, so it shouldn't cause the creation of statues.