Don't let pit fiend create pit/hole on stairs, ladder, altar, or magic chest
This could only happen if the defender was standing directly on the
terrain in question, but still, the game behaves very strangely and
somewhat inconsistently if any kind of trap is on those terrains,
so I think it's best to forbid it altogether.  Unfortunately this
makes standing on the stairs a reliable strategy for blocking a pit
fiend's primary attack, which is not ideal.  Perhaps later I can give
them the ability to knock you down the down stairs, or horizontally
off the up stairs.  But that can be for another commit, another day.
This commit is just the bugfix.