Bugfix: enchanting dragon scales onto yourself
Previously, due to an oversight in the implementation, enchanting
dragon scales onto yourself (without any body armor) allowed you to
polymorph into any monster you wanted, if you had polymorph control.
This is wrong; the scroll of enchant armor is not meant to be a
general-purpose source of polymorph; the true source of polymorph
in that scenario is the scales, and they can only polymorph you
into one thing:  a dragon.