Bugfix: jitterbug displacement and pits.
If a jitterbug displaces you out of a pit, you are now no longer
trapped in the pit.  While I was at it, I implemented the TODO
about getting jitterbug-displaced _into_ a trap; this now activates
the trap in the usual expected way.

Also, correctly document the criterion for unlocking Priest role.