Track max alignment so far in this game; mood ring enhancement.
The game now tracks what the highest alignment is that the player
has achieved so far in the game, and reports in the history the
first time the player reaches certain levels.  The first time you
become pious, it gets livelogged.

The mood ring is now slightly more useful, by dint of having more
colors and therefore more nuance.  Notably, there is now a color
that specifically indicates that you are pious.  Also, the ring
now distinguishes between zero and negative alignment.