Setseed logging and alignment record reporting fixes.
In setseed games, the dumplog and xlogfile should now report the
seed that was set.

Don't say things like "Your conscience bothers you a little less"
when your alignment record is as good as it has ever been.  This
was a bug introduced in Fourk, probably as a side-effect of some
changes to the initial setup of the player's alignment.  Notably,
starting alignment depends on your alignment in Fourk, i.e., your
numeric alignment at game start is dependent on whether you are
lawful, neutral, or chaotic; the code surrounding the messages
about your conscience, was not taking this into account, because
said code was inherited from vanilla and never updated when the
alignment-dependent initial alignment was introduced.  Since we
are now tracking your character's personal "historic maximum
alignment", however, we can simplify the logic anyway.