Don't replace polycontrol ring with booby prizes
A player may unknowingly be wearing an amulet of change simply because it's an uncursed amulet; that early in the game, it's not a clear signal that the player doesn't want polycontrol.

If they do want polycontrol, they probably don't need an amulet of flying, as flight can be achieved more affordably by polymorphing.

Putting a non-ring in the bag of rings is going to feel random to unspoiled players, buggy to lightly-spoiled players, and leak a very arcane piece of information to a source-diver who hasn't identified the ring(s) in their possession.

Instead, if the player doesn't need the coveted ring of foo_control, give them the counterpart, a ring of foo. If they don't need that ring either, give a random ring from a range that is unlikely to result in duplicate rings.